Add error handling to implemented calls

+++ service/storage.js
@@ -1,3 +1,5 @@
+const uuid = require('uuid/v4')
+const BadRequestError = require('../error/bad-request')
 const client = require('../client/firestore')
 
 exports.find = async payload => {
@@ -6,7 +8,9 @@ exports.find = async payload => {
 
 exports.get = async payload => {
     const { collectionName, itemId } = payload
-    //TODO add request payload validation
+    if (!collectionName) throw new BadRequestError('Missing collectionName in request body')
+    if (!itemId) throw new BadRequestError('Missing itemId in request body')
+
     const document = await client.get(collectionName, itemId)
 
     if (!document.exists) {
@@ -21,27 +25,31 @@ exports.get = async payload => {
 
 exports.insert = async payload => {
     const { collectionName, item } = payload
+    if (!collectionName) throw new BadRequestError('Missing collectionName in request body')
+    if (!item) throw new BadRequestError('Missing item in request body')
+    
+    if (!item._id) item._id = uuid()
     await client.insert(collectionName, item)
-    //TODO handle already existing item case
-    //TODO handle missing _id case (autogeneration)
 
     return item
 }
 
 exports.update = async payload => {
     const { collectionName, item } = payload
-    await client.update(collectionName, item)
+    if (!collectionName) throw new BadRequestError('Missing collectionName in request body')
+    if (!item) throw new BadRequestError('Missing item in request body')
 
-    //TODO handle non existing item case (not upsert)
+    await client.update(collectionName, item)
 
     return item
 }
 
 exports.remove = async payload => {
     const { collectionName, itemId } = payload
-    //TODO add request payload validation
+    if (!collectionName) throw new BadRequestError('Missing collectionName in request body')
+    if (!itemId) throw new BadRequestError('Missing itemId in request body')
+
     await client.delete(collectionName, itemId)
-    //TODO handle nonexistent case
 
     return {
         _id: itemId
